requirements.txt
# Django core
Django>=5.1.3,<6.0

# Environment variables
python-decouple>=3.8

# Django REST Framework
djangorestframework>=3.20

# JWT Authentication
djangorestframework-simplejwt>=6.0

# Redis support
django-redis>=6.1

# Django RQ for background jobs
django-rq>=3.1

# Redis client
redis>=5.0

# Email sending (optional, for Gmail/SMTP)
# smtplib is built-in, no need to install separately

# UUID support (built-in in Python >=3.6)
# No external package needed
